app/requirements.txt
langchain-core
langgraph
langchain-google-vertexai
google-generativeai
python-dotenv
google-auth
google-cloud-aiplatform
